partner_chains_node_commands

Trait PartnerChainRuntime

pub trait PartnerChainRuntime {
    type Keys: MaybeFromCandidateKeys;

    // Required methods
    fn create_chain_spec(config: &CreateChainSpecConfig<Self::Keys>) -> Value;
    fn key_definitions() -> Vec<KeyDefinition<'static>>;
}

Trait defining Partner Chain governance related types.

Required Associated Types

type Keys: MaybeFromCandidateKeys

User keys type

Required Methods

fn create_chain_spec(config: &CreateChainSpecConfig<Self::Keys>) -> Value

User defined function to create a chain spec given the PC configuration

fn key_definitions() -> Vec<KeyDefinition<'static>>

Names and schemes of keys used by the runtime.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.